pyproject.toml
[project]
name = "pdf-page-images"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-docx>=1.1.2",
    "typer>=0.15.2",
]
